removed spurious comment
cdent says:
"""
I left out the "for python 2.4 compatibility" thing on purpose, for the
simple reason that urllib2 does not provide quote itself, urllib does.
That it is in urllib2's package space is accidental. If you look at the
docs for urllib2 it makes no mention of quote() so it seems more
accurate to actually just use urllib, whether trying to support 2.4 or
not.
"""
http://github.com/tiddlyweb/tiddlywebplugins.instancer/commit/7c0d4f44270efe069c6eaf771e59986ff807c357#comments