When bag names are assigned to the 'bag' attribute on a tiddler, make it unicode.
This is necessary because all tiddler attribute need to be unicode
in various aspects of the system, but notably when being indexed by
whoosher. The Tiddler class intentionally does not manage the attributes
to check for this to keep it as lightweight and transparent as possible.

The problem has not shown up before because most of the time Tiddler
objects are created as Tiddler(title, bag_name) and __init__ does
handle ensuring unicode.